+++ python/rational_linkages/MotionDesigner.py
@@ -1,19 +1,24 @@
-from typing import Union
-import numpy as np
 import sys
+from typing import Union
+
+import numpy as np
+import pyqtgraph.opengl as gl
 
 # PyQt and Pyqtgraph imports
 from PyQt6 import QtCore, QtWidgets
-import pyqtgraph.opengl as gl
 
 # Import your custom classes (adjust the import paths as needed)
 from .DualQuaternion import DualQuaternion
 from .MotionInterpolation import MotionInterpolation
-from .RationalMechanism import RationalMechanism
-from .RationalCurve import RationalCurve
+from .PlotterPyqtgraph import (
+    FramePlotHelper,
+    InteractivePlotterWidget,
+    PlotterPyqtgraph,
+)
 from .PointHomogeneous import PointHomogeneous
+from .RationalCurve import RationalCurve
+from .RationalMechanism import RationalMechanism
 from .TransfMatrix import TransfMatrix
-from .PlotterPyqtgraph import PlotterPyqtgraph, FramePlotHelper, InteractivePlotterWidget
 
 
 class MotionDesigner:
@@ -77,7 +82,7 @@ class MotionDesigner:
     """
     def __init__(self,
                  method: str,
-                 initial_points_or_poses: list[PointHomogeneous, DualQuaternion] = None,
+                 initial_points_or_poses: list[Union[PointHomogeneous, DualQuaternion]] = None,
                  arrows_length: float = 1.0,
                  white_background: bool = False):
         """